Resolve problems with merge

+++ src/main/java/com/crowd/service/ProjectService.java
@@ -45,7 +45,10 @@ public class ProjectService {
     }
 
     public void delete(Project project) {
-        projectDAO.delete(project);
+        if (project.getStatus().equals(Status.NOT_STARTED) ||  project.getStatus().equals(Status.IN_PROGRESS)){
+            projectDAO.delete(project);
+        }
+
     }
 
     public List<Project> findByPartOfProjectName(String projectName) {
@@ -114,9 +117,9 @@ public class ProjectService {
                 }
                 update(project);
             }
-            break;
-            default:
                 break;
+
+            default: break;
         }
     }
